# docs/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: A robust RESTful API for a virtual garden management game built with
    Go.
  title: Virtual Garden Management System API
  version: "1.0"
paths:
  /audit/gardens/{garden_id}/activity:
    get:
      consumes:
      - application/json
      description: Get recent activity for a specific garden
      parameters:
      - description: Garden ID
        in: path
        name: garden_id
        required: true
        type: string
      - default: 20
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Garden activity logs
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Get garden activity
      tags:
      - audit
  /audit/logs:
    get:
      consumes:
      - application/json
      description: Get audit logs with optional filtering
      parameters:
      - description: User ID filter
        in: query
        name: user_id
        type: string
      - description: Action filter
        in: query
        name: action
        type: string
      - description: Resource filter
        in: query
        name: resource
        type: string
      - description: Resource ID filter
        in: query
        name: resource_id
        type: string
      - description: Status filter
        in: query
        name: status
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Audit logs with pagination
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Get audit logs
      tags:
      - audit
  /audit/stats:
    get:
      consumes:
      - application/json
      description: Get audit statistics for the system
      parameters:
      - default: 30
        description: Number of days to analyze
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Audit statistics
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Get audit statistics
      tags:
      - audit
  /audit/users/{user_id}/activity:
    get:
      consumes:
      - application/json
      description: Get recent activity for a specific user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - default: 20
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User activity logs
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Get user activity
      tags:
      - audit
...
securityDefinitions:
  bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"